+++ microhapulator/tests/test_pipe.py
@@ -39,9 +39,6 @@ def test_pipe_missing_files(tmp_path):
 
 
 def test_pipe_gbr_usc10(tmp_path):
-    hg38 = str(tmp_path / "hg38-placeholder.fasta")
-    copyfile(data_file("refr/usc10-refr.fna"), hg38)
-    run(["bwa", "index", hg38])
     arglist = [
         "pipe",
         data_file("refr/usc10-refr.fna"),
@@ -53,8 +50,6 @@ def test_pipe_gbr_usc10(tmp_path):
         "--threads=1",
         "--static=5",
         "--dynamic=0.02",
-        "--hg38",
-        hg38,
     ]
     args = microhapulator.cli.get_parser().parse_args(arglist)
     microhapulator.cli.pipe.main(args)